Submit to git early and often
Wait way too long to commit to git. I don't
even know where to begin commenting on the what or why. To be fair this
is my first project working with git and I am still getting used to
committing early and often.

This commit message is for my future self to reference how far I have
come with git.